added songs and tests

+++ generate_compression.py
@@ -3,10 +3,10 @@ import gzip
 import bz2
 import lzma
 from typing import Optional, List, Tuple
-
 import zstandard as zstd
 import click
 import shutil
+import csv
 
 
 class Compressor:
@@ -123,13 +123,30 @@ class FileManager:
                 print(f"Directory does not exist: {directory}")
 
 
-
 class CompressionInstance:
-    def __init__(self, compression_methods: List[str]):
+    def __init__(self, compression_methods: List[str], \
+        folder_test: Optional[str] = None, file_test: Optional[str] = None, \
+            noise_type: Optional[str] = None, noise_percentage: Optional[str] = None, \
+            test_start: Optional[str] = None, test_duration: Optional[str] = None, test_percentage: Optional[str] = None, \
+            csv_file: Optional[str] = None):
+        folder_test = folder_test if folder_test is not None else "interpretations_trim"
+        file_test = file_test if file_test is not None else "output"
+        
+        self.file_test = file_test
+        
+        self.noise_type = noise_type if noise_type is not None else "none"
+        self.noise_percentage = noise_percentage if noise_percentage is not None else "none"
+        
+        self.test_percentage = test_percentage if test_percentage is not None else "none"
+        self.test_start = test_start if test_start is not None else "none"
+        self.test_duration = test_duration if test_duration is not None else "none"
+        
+        self.csv_file = csv_file
+        self.folder_test = folder_test
         self.compression_methods = compression_methods
         self.signatures_dir = './song_signatures'
         self.compressed_dir = './compressed'
-        self.trim_segment_file = './trim/output_segment'
+        self.trim_segment_file = f'./{folder_test}/{file_test}'
 
     def compress_files(self) -> None:
         """
@@ -159,10 +176,12 @@ class CompressionInstance:
         for method in self.compression_methods:
             compressed_trim_segment_file = f'{self.trim_segment_file}.{method}'
             Compressor.compress_file(method, self.trim_segment_file, compressed_trim_segment_file)
-            compressed_files.append((method, compressed_trim_segment_file, FileManager.get_file_size_in_bits(compressed_trim_segment_file)))
+            file_size_bits = FileManager.get_file_size_in_bits(compressed_trim_segment_file)
+            if file_size_bits is not None:
+                compressed_files.append((method, compressed_trim_segment_file, file_size_bits))
         return compressed_files
 
-    def calculate_normalized_compressed_distance(self, compressed_files: List[Tuple[str, str, int]]) -> None:
+    def calculate_normalized_compressed_distance(self, compressed_files: List[Tuple[str, str, int]], folder_test: Optional[str] = None) -> None:
         """
         Calculates the normalized compressed distance for each signature file against the compressed trim segment files.
 
@@ -170,6 +189,15 @@ class CompressionInstance:
         :return: None
         """
         files_ordered_by_bits = []
+        csv_file = getattr(self, 'csv_file', None)
+
+        csv_writer = None
+        if csv_file:
+            file_exists = os.path.isfile(csv_file)
+            csv_file_handle = open(csv_file, mode='a', newline='')
+            csv_writer = csv.writer(csv_file_handle)
+            if not file_exists:
+                csv_writer.writerow(['file', 'NCD', 'encoding', 'noise_type', 'file_tested', 'noise_percentage', 'test_start', 'test_duration', 'test_percentage'])
 
         for method, compressed_trim_segment_file, compressed_trim_segment in compressed_files:
             for signature_file in os.listdir(self.signatures_dir):
@@ -179,27 +207,44 @@ class CompressionInstance:
 
                 if os.path.exists(compressed_file_path):
                     file_size_bits = FileManager.get_file_size_in_bits(compressed_file_path)
-                    combined_file = './trim/temp_combined'
+                    
+                    folder_test = folder_test if folder_test is not None else "interpretations_trim"
+                    combined_file = f'./{folder_test}/temp_combined'
+                    
                     FileManager.concatenate_files(signature_file_path, self.trim_segment_file, combined_file)
                     Compressor.compress_file(method, combined_file, f'{combined_file}.{method}')
                     combined_file_size_bits = FileManager.get_file_size_in_bits(f'{combined_file}.{method}')
+                    
+                    if combined_file_size_bits is None:
+                        continue
+                    
                     normalized_compressed_distance = (combined_file_size_bits - min(file_size_bits, compressed_trim_segment)) / max(file_size_bits, compressed_trim_segment)
-                    files_ordered_by_bits.append((compressed_file_path, normalized_compressed_distance))
+                    result = (base_name, normalized_compressed_distance, method, self.noise_type, self.file_test, self.noise_percentage, self.test_start, self.test_duration, self.test_percentage)
+                    files_ordered_by_bits.append(result)
+                    
+                    if csv_writer:
+                        csv_writer.writerow(result)
+                    
                     os.remove(f'{combined_file}.{method}')
                     os.remove(combined_file)
 
-        files_ordered_by_bits.sort(key=lambda item: item[1])
-        list(map(lambda item: print(f"Path: {item[0]}, Normalized Compressed Distance: {item[1]}"), files_ordered_by_bits))
+        if csv_writer:
+            csv_file_handle.close()
+        else:
+            files_ordered_by_bits.sort(key=lambda item: item[1])
+            for item in files_ordered_by_bits:
+                print(f"Path: {item[0]}, Normalized Compressed Distance: {item[1]}, Encoding: {item[2]}, Noise: {item[3]}")
 
-    def run(self) -> None:
+    def run(self, compress) -> None:
         """
         Runs the compression application.
 
         :return: None
         """
-        self.compress_files()
+        if compress:
+            self.compress_files()
         compressed_files = self.compress_trim_segment()
-        self.calculate_normalized_compressed_distance(compressed_files)
+        self.calculate_normalized_compressed_distance(compressed_files, self.folder_test)
 
 
 @click.command()
@@ -208,7 +253,15 @@ class CompressionInstance:
 @click.option('-bzip2', is_flag=True, help='Compress using bzip2.')
 @click.option('-lzma', is_flag=True, help='Compress using lzma.')
 @click.option('-zstd', is_flag=True, help='Compress using zstd.')
-def main(compress, gzip, bzip2, lzma, zstd):
+@click.option('-folder-test', help="Folder where tests are.")
+@click.option('-test-file', help="Signature file to test.")
+@click.option('-noise-type', help="Type of noise used.")
+@click.option('-noise-percentage', help="Percentage of noise used.")
+@click.option('-test-start', help="Test file start time.")
+@click.option('-test-duration', help="Test file duration.")
+@click.option('-test-percentage', help="Percentage of considered audio, compared to the original song")
+@click.option('-csv-file', help="File to write the tests to.")
+def main(compress, gzip, bzip2, lzma, zstd, folder_test, test_file, noise_type, noise_percentage, test_start, test_duration, test_percentage, csv_file):
     """
     Main function to compress songs using various compression methods.
 
@@ -217,21 +270,25 @@ def main(compress, gzip, bzip2, lzma, zstd):
     :param bzip2: Whether to compress using bzip2.
     :param lzma: Whether to compress using lzma.
     :param zstd: Whether to compress using zstd.
+    :param folder-test: Folder where the files to be tested are located.
+    :param test-file: File to be tested.
+    :param noise-type: Type of noise used.
+    :param noise-percentage: Percentage of noise used.
+    :param test-start: Test file start time.
+    :param test-duration: Test file duration.
+    :param test-percentage: Test file duration.
+    :param csv-file: File to write the tests to.
     :return: None
     """
 
-    if not compress:
-        FileManager.list_existing_files()
+    compress_flags = {'gzip': gzip, 'bzip2': bzip2, 'lzma': lzma, 'zstd': zstd}
+    selected_methods = [method for method, selected in compress_flags.items() if selected]
+    if selected_methods:
+        app = CompressionInstance(selected_methods, folder_test, test_file, noise_type, noise_percentage, test_start, test_duration, test_percentage, csv_file)
+        app.run(compress)
     else:
-        compress_flags = {'gzip': gzip, 'bzip2': bzip2, 'lzma': lzma, 'zstd': zstd}
-        selected_methods = [method for method, selected in compress_flags.items() if selected]
-        if selected_methods:
-            app = CompressionInstance(selected_methods)
-            app.run()
-        else:
-            print("No valid compression method specified.")
+        print("No valid compression method specified.")
 
 
 if __name__ == "__main__":
     main()
-
